Fix CFM-05 issue from the audit
The audit showed that there is an unreachable if branch in the
listenForRead() method: In cases where s == connectivity.Ready is false
and the else block is executed, the "if done != nil" block is never
actually executed. This is because the "done" variable is always nil.

Note that this change will end up closing _all_ goroutines that are running
the handleRead() method. This makes sense, espcially if the connection
to the server is severed.