Configure the call timeout by providing a deadline/timeout in the provided context
We had previously hardcoded a timeout value for a client to wait for an RPC response
to 2 seconds. This has now been changed to respect the context that is already provided,
and will instead timeout when the context is cancelled.

By default we now have now deadline for a q request to complete unless explicitly set in
the context