Fixes an issue where clients were unable to reconnect
When dealing with clients with poor networks (high latency or unstable
connectivity) the server would occassionally not drop the connection,
leading to clients being unable to reconnect because the server only
allows a single connection per Public Key.

This change makes the server send periodic pings to the clients, and
will drop the connection if no response is received. This allows the
client to re-obtain a connection when the network comes back up for them
again.

The ping handler has also been removed in favour of the default
implementation which was performing the same function but with
a lower ping read timeout.